Fix `BsonInvalidOperationException` when current type not read
A `BsonReader` maintains the current Bson type it has read. But it returns
`null` if the type was not already read. The user must ensure to first call
`readBsonType` prior to calling `getCurrentBsonType`.

```
org.bson.BsonInvalidOperationException: readUndefined can only be called when CurrentBSONType is UNDEFINED, not when CurrentBSONType is
INT64. (AbstractBsonReader.java:690)
  org.bson.AbstractBsonReader.verifyBSONType(AbstractBsonReader.java:690)
  org.bson.AbstractBsonReader.checkPreconditions(AbstractBsonReader.java:722)
  org.bson.AbstractBsonReader.readUndefined(AbstractBsonReader.java:478)
  reactivemongo.api.bson.msb.BSONDecoder.decode(HandlerConverters.scala:271)
  reactivemongo.api.bson.msb.BSONDecoder.decode$(HandlerConverters.scala:201)
  reactivemongo.api.bson.msb.HandlerConverters$DefaultCodec.decode(HandlerConverters.scala:137)
  reactivemongo.HandlerConverterSpec.$anonfun$new$7(HandlerConverterSpec.scala:83)
```